Fix SonarCloud bugs, blockers, and string duplication
Bugs (integer overflow in FFM code):
- Use long arithmetic for MemorySegment offsets in
  LogicalProcessorInformationFFM (13 issues)
- Use long literals in SessionWtsDataFFM constant expressions
- Use long offset in WindowsCentralProcessorFFM frequency parsing
- Use long in WindowsNetworkIfFFM alias slice size
- Use long in Advapi32UtilFFM buffer allocation

Blockers (missing test assertions):
- Add assertDoesNotThrow to IOReportClientTest.testCloseIsIdempotent
- Add futures size assertion to GpuStatsTest.testConcurrentAccessDoesNotThrow

String duplication (S1192):
- Extract Temperature/Voltage constants in WindowsSensorsFFM
- Extract GPU_CORE constant in WindowsGpuStatsFFM
- Extract WMI_HANDLER_NULL constant in Win32ProcessorFFM and
  Win32ProcessorJNA

+++ oshi-core-ffm/src/main/java/oshi/hardware/platform/windows/WindowsCentralProcessorFFM.java
@@ -122,7 +122,7 @@ final class WindowsCentralProcessorFFM extends WindowsCentralProcessor {
                 MemorySegment openedKey = hKey.get(ValueLayout.ADDRESS, 0);
                 try {
                     // Get first subkey name
-                    MemorySegment nameBuffer = arena.allocate(256 * 2); // WCHAR[256]
+                    MemorySegment nameBuffer = arena.allocate(256 * 2L); // WCHAR[256]
                     MemorySegment nameLen = arena.allocate(ValueLayout.JAVA_INT);
                     nameLen.set(ValueLayout.JAVA_INT, 0, 256);
                     if (Advapi32FFM.RegEnumKeyEx(openedKey, 0, nameBuffer, nameLen, MemorySegment.NULL,
@@ -277,7 +277,7 @@ final class WindowsCentralProcessorFFM extends WindowsCentralProcessor {
                 return freqs;
             }
             for (int i = 0; i < freqs.length; i++) {
-                int offset = i * PPI_SIZE;
+                long offset = i * (long) PPI_SIZE;
                 // ProcessorPowerInformation: number(4), maxMhz(4), currentMhz(4), mhzLimit(4), maxIdleState(4),
                 // currentIdleState(4)
                 if (fieldIndex == 1) {
